Practice/django-image-upload/requirements.txt
Django
djangorestframework
Pillow
drf-yasg
```